refactor secrets plugin
